initial posteriorplot for bokeh
